third_party/blink/web_tests/platform/generic/fast/forms/radio-checkbox-restore-indeterminate-expected.txt

A B
Intermediate A Intermediate B
C D
Click did not unselect all radio buttons: PASS

Checking whether default-prevented click clobbered indeterminate state:
On checked radio button: PASS
On unchecked radio button: PASS
On checked checkbox: PASS
On unchecked checkbox: PASS

Checking whether default-prevented click clobbered checked state:
On checked radio button: PASS
On unchecked radio button: PASS
On checked checkbox: PASS
On unchecked checkbox: PASS
